Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -73,5 +73,5 @@ total_age = 0
 
 end 
 binding.pry
-return total_age.float()/contestant_counter
+return total_age/contestant_counter
 end
